requirements.txt
astroid==2.4.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
google-api-core==1.25.0
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
h11==0.11.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libcst==0.3.16
mccabe==0.6.1
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.3
pylint==2.6.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
uvicorn==0.13.2
wrapt==1.12.1